+++ setup.py
@@ -71,7 +71,7 @@ setup(name='mcmcsampler',
       description=DOCLINES[0],
       long_description="\n".join(DOCLINES[2:]),
       version=__version__,
-      url='https://github.com/choderalab/perses',
+      url='https://github.com/choderalab/openmm-mcmc',
       platforms=['Linux', 'Mac OS-X', 'Unix'],
       classifiers=CLASSIFIERS.splitlines(),
       packages=['mcmcsampler', 'mcmcsampler.tests'],
